feat: integrate useClient hook
