+++ src/features/auth/hooks/use-register-form.ts
@@ -1,5 +1,4 @@
 import { zodResolver } from '@hookform/resolvers/zod';
-import { useNavigate } from '@tanstack/react-router';
 import { AxiosError } from 'axios';
 import { TFunction } from 'i18next';
 import { isValidPhoneNumber } from 'libphonenumber-js';
@@ -46,7 +45,6 @@ export const createSchema = (t: TFunction) =>
 export type RegisterFormSchema = z.infer<ReturnType<typeof createSchema>>;
 
 export const useRegisterForm = () => {
-  const navigate = useNavigate({ from: '/auth/register' });
   const { t } = useTranslation();
   const form = useForm({
     resolver: zodResolver(createSchema(t)),
@@ -108,7 +106,6 @@ export const useRegisterForm = () => {
   const onSubmit = form.handleSubmit(async (data: RegisterFormSchema) => {
     try {
       await register(data);
-      navigate({ to: '/auth/register/done' });
     } catch (err) {
       if (err instanceof AxiosError) {
         switch (err.response?.status) {
